+++ config/chain/config.go
@@ -11,13 +11,13 @@ import (
 )
 
 type GeneralChainConfig struct {
-	Name           string `mapstructure:"name"`
-	Id             *uint8 `mapstructure:"id"`
-	Endpoint       string `mapstructure:"endpoint"`
-	Type           string `mapstructure:"type"`
-	BlockstorePath string `mapstructure:"blockstorePath"`
-	FreshStart     bool   `mapstructure:"fresh"`
-	LatestBlock    bool   `mapstructure:"latest"`
+	Name           string  `mapstructure:"name"`
+	Id             *uint64 `mapstructure:"id"`
+	Endpoint       string  `mapstructure:"endpoint"`
+	Type           string  `mapstructure:"type"`
+	BlockstorePath string  `mapstructure:"blockstorePath"`
+	FreshStart     bool    `mapstructure:"fresh"`
+	LatestBlock    bool    `mapstructure:"latest"`
 	Key            string
 	Insecure       bool
 }
